+++ CHANGELOG.md
@@ -4,10 +4,9 @@ All notable changes to this project will be documented in this file. See [standa
 
 ## [1.2.0](https://github.com/routelink/client/compare/v1.1.0...v1.2.0) (2024-04-28)
 
-
 ### Features
 
-* add mobx to Transport component ([#104](https://github.com/routelink/client/issues/104)) ([addae22](https://github.com/routelink/client/commit/addae2246ad9d9233284965ad5147661fd0ab3f7))
+- add mobx to Transport component ([#104](https://github.com/routelink/client/issues/104)) ([addae22](https://github.com/routelink/client/commit/addae2246ad9d9233284965ad5147661fd0ab3f7))
 
 ## [1.1.0](https://github.com/routelink/client/compare/v1.0.0...v1.1.0) (2024-04-23)